\name{modelULPresence}
\alias{modelULPresence}
\docType{data}
\title{
  Registered Models for the Grid Cell Search Data
}
\description{
  This dataset contains the JAGS code for the registered model.
}
\usage{
data("modelULPresence")
}
\format{
  The format is a character string containg the JAGS code, directly
  usable with JAGS.
}
\details{
  
  This dataset contain the JAGS code for the model used to describe the
  distribution of unknown leks in the grid cells, as a function of the
  proportion of the cell covered by the area of presence of the
  capercaillie, of the presence of already known leks, and of the
  geographic region.
  
}
\source{
  Calenge C., Menoni E., Milhau B., Montadert M. (in
  prep.). Building a Capercaillie Monitoring program in the Pyrenees 
  Mountains France based on Participatory Research.
}
\references{
  Calenge C., Menoni E., Milhau B., Montadert M. (in
  prep.). Building a Capercaillie Monitoring program in the Pyrenees 
  Mountains France based on Participatory Research.
}
\examples{
modelULPresence

}
\keyword{datasets}
